# Conflicts:
#	pom.xml
#	src/main/java/ru/yandex/practicum/filmorate/controller/Controller.java
#	src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
#	src/main/java/ru/yandex/practicum/filmorate/controller/IController.java
#	src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
#	src/main/java/ru/yandex/practicum/filmorate/exception/FilmorateException.java
#	src/main/java/ru/yandex/practicum/filmorate/exception/ValidationException.java
#	src/main/java/ru/yandex/practicum/filmorate/model/Film.java
#	src/main/java/ru/yandex/practicum/filmorate/model/FilmorateObject.java
#	src/main/java/ru/yandex/practicum/filmorate/model/User.java
#	src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
#	src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
#	src/main/java/ru/yandex/practicum/filmorate/validator/NotContainSpace.java
#	src/main/java/ru/yandex/practicum/filmorate/validator/NotContainSpaceValidator.java
#	src/test/java/ru/yandex/practicum/filmorate/controller/ControllerTest.java
#	src/test/java/ru/yandex/practicum/filmorate/controller/FilmControllerTest.java
#	src/test/java/ru/yandex/practicum/filmorate/controller/TestData.java
#	src/test/java/ru/yandex/practicum/filmorate/controller/UserControllerTest.java
#	src/test/java/ru/yandex/practicum/filmorate/tools/GsonAdapter.java

+++ src/main/java/ru/yandex/practicum/filmorate/model/FilmorateObject.java
@@ -1,13 +1,10 @@
 package ru.yandex.practicum.filmorate.model;
 
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 @Getter
 @Setter
 @ToString
 public abstract class FilmorateObject {
-    protected int id;
+    protected long id;
 }
